fix: M8 audit fix pass — H1 + all MEDIUMs, quick LOWs, backlog recorded
HIGH: link-preview panel raised above the profile modal (z-46/47) — chips
in profile bodies opened it invisibly and its capture-phase Escape ate the
modal's. MEDIUMs: matcher scores 'Bi - male/female preference' (match on
preferred gender, weak otherwise, substring trap tested); insights/history
loads catch failures (error state + Retry on the tab); history-rail remove
is a real sibling button (valid ARIA, keyboard-reachable); loadProfile
guards superseded responses from clobbering a fresh refresh; server-side
in-flight dedup coalesces concurrent character loads onto one budget unit;
cross-identity profile-cache sharing documented in self-hosting.md.
Quick LOWs: img referrerPolicy no-referrer, image-extension allowlist,
DETACHED_DISCONNECT_HOURS .int(), index-refresh comment corrected.
Remaining LOWs recorded as the M8 audit backlog in milestones.md.

+++ apps/server/src/config.ts
@@ -95,7 +95,7 @@ const configSchema = z.object({
    * The in-memory vault keeps the credentials, so the next attach
    * reconnects automatically with the exact channel set.
    */
-  DETACHED_DISCONNECT_HOURS: z.coerce.number().min(0).default(72),
+  DETACHED_DISCONNECT_HOURS: z.coerce.number().int().min(0).default(72),
   /**
    * Character-data-class requests allowed per sliding hour (M8 profiles).
    * F-List's published limit is 200/hour; the 170 default leaves headroom —
